+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed a bug with `FromParams` that caused `.from_params()` to fail when the params contained
+  an object that was already instantiated.
+
 ## [v0.3.1](https://github.com/allenai/tango/releases/tag/v0.3.1) - 2021-10-29
 
 ### Changed